added custom boss to GY graveyard
hey Honey. So here is my idea for a custom boss. I did it quite quickly and maybe a bit crudely. I tried doing it for level 70 and might have messed up some things in the process. I was just curious about the model of the boss itself and the adds. Does this have to be done manually? Cheers in advance for the help!